Rename all reducers files to singular form
