+++ src/auth/__tests__/authReducers.js
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 import * as authTypes from '../authActions';
-import authReducers from '../authReducers';
+import authReducer from '../authReducer';
 
 const initialStateMock = {
   isAuthenticated: false,
@@ -12,20 +12,20 @@ const initialStateMock = {
 
 describe('authReducer', () => {
   it('is expected to return an object', () => {
-    expect(authReducers(undefined, {})).to.be.an('object');
+    expect(authReducer(undefined, {})).to.be.an('object');
   });
   it('is expected to return initial state on undefined action', () => {
-    expect(authReducers(undefined, {})).to.deep.equal(initialStateMock);
+    expect(authReducer(undefined, {})).to.deep.equal(initialStateMock);
   });
 
   describe(`dispatching action ${authTypes.LOGIN_REQUEST}`, () => {
     it('is expected to return isFetching as true', () => {
-      expect(authReducers(undefined, { type: authTypes.LOGIN_REQUEST })).to.have
+      expect(authReducer(undefined, { type: authTypes.LOGIN_REQUEST })).to.have
         .property('isFetching')
         .and.equal(true);
     });
     it('is expected to return isAuthenticated as false', () => {
-      expect(authReducers(undefined, { type: authTypes.LOGIN_REQUEST })).to.have
+      expect(authReducer(undefined, { type: authTypes.LOGIN_REQUEST })).to.have
         .property('isAuthenticated')
         .and.equal(false);
     });
@@ -33,17 +33,17 @@ describe('authReducer', () => {
 
   describe(`dispatching action ${authTypes.LOGIN_SUCCESS}`, () => {
     it('is expected to return isFetching as false', () => {
-      expect(authReducers(undefined, { type: authTypes.LOGIN_SUCCESS })).to.have
+      expect(authReducer(undefined, { type: authTypes.LOGIN_SUCCESS })).to.have
         .property('isFetching')
         .and.equal(false);
     });
     it('is expected to return isAuthenticated as true', () => {
-      expect(authReducers(undefined, { type: authTypes.LOGIN_SUCCESS })).to.have
+      expect(authReducer(undefined, { type: authTypes.LOGIN_SUCCESS })).to.have
         .property('isAuthenticated')
         .and.equal(true);
     });
     it('is expected to return user from the action', () => {
-      expect(authReducers(undefined, { type: authTypes.LOGIN_SUCCESS, user: 'testUser' })).to.have
+      expect(authReducer(undefined, { type: authTypes.LOGIN_SUCCESS, user: 'testUser' })).to.have
         .property('user')
         .and.equal('testUser');
     });
@@ -51,18 +51,18 @@ describe('authReducer', () => {
 
   describe(`dispatching action ${authTypes.LOGIN_FAILURE}`, () => {
     it('is expected to return isFetching as false', () => {
-      expect(authReducers(undefined, { type: authTypes.LOGIN_FAILURE })).to.have
+      expect(authReducer(undefined, { type: authTypes.LOGIN_FAILURE })).to.have
         .property('isFetching')
         .and.equal(false);
     });
     it('is expected to return isAuthenticated as false', () => {
-      expect(authReducers(undefined, { type: authTypes.LOGIN_FAILURE })).to.have
+      expect(authReducer(undefined, { type: authTypes.LOGIN_FAILURE })).to.have
         .property('isAuthenticated')
         .and.equal(false);
     });
     it('is expected to return errorMessage from action.message', () => {
       expect(
-        authReducers(undefined, { type: authTypes.LOGIN_FAILURE, message: 'testMessage' }),
+        authReducer(undefined, { type: authTypes.LOGIN_FAILURE, message: 'testMessage' }),
       ).to.have
         .property('errorMessage')
         .and.equal('testMessage');
@@ -71,7 +71,7 @@ describe('authReducer', () => {
 
   describe(`dispatching action ${authTypes.LOGOUT_SUCCESS}`, () => {
     it('is expected to return isAuthenticated as false', () => {
-      expect(authReducers(undefined, { type: authTypes.LOGOUT_SUCCESS })).to.have
+      expect(authReducer(undefined, { type: authTypes.LOGOUT_SUCCESS })).to.have
         .property('isAuthenticated')
         .and.equal(false);
     });
