+++ src/favorites/favoritesReducer.js
